# openapi/notes-api-v1.yaml
...
openapi: 3.0.0
info:
  title: OSM Notes API
  version: '0.1.0'
  description: |
    REST API for OSM Notes Analytics and Ingestion. Provides programmatic access to
    user profiles, country analytics, notes, and real-time metrics.

    Responses that contain data from OpenStreetMap or derivatives include an
    `attribution` field (text and url) so clients can display proper attribution
    (e.g. "© OpenStreetMap contributors").
  contact:
    name: OSM Notes Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://notes-api.osm.lat
    description: Production server
security:
  - UserAgent: []
components:
  securitySchemes:
    UserAgent:
      type: apiKey
      in: header
      name: User-Agent
      description: 'User-Agent header required for all requests. Format: AppName/Version (Contact)'
  schemas:
    Error:
      type: object
      required: [error, message, statusCode]
      properties:
        error:
          type: string
          description: Error type name
          example: Bad Request
        message:
          type: string
          description: Human-readable error message
          example: Invalid note ID
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
...
    Note:
      type: object
      required: [note_id, latitude, longitude, status, created_at]
      properties:
        note_id:
          type: integer
          description: OSM note ID
          example: 12345
        latitude:
          type: number
          format: float
          description: Latitude coordinate
          example: 4.6097
        longitude:
          type: number
          format: float
          description: Longitude coordinate
          example: -74.0817
        status:
          type: string
          enum: [open, closed, reopened]
          description: Note status
          example: open
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: '2024-01-15T10:30:00Z'
        closed_at:
          type: string
          format: date-time
          nullable: true
          description: Closing timestamp
        id_user:
          type: integer
          nullable: true
          example: 67890
        id_country:
          type: integer
          nullable: true
          example: 42
        comments_count:
          type: integer
          example: 3
    NoteComment:
      type: object
      properties:
        comment_id:
          type: integer
          example: 1
        note_id:
          type: integer
          example: 12345
        user_id:
          type: integer
          nullable: true
          example: 67890
        username:
          type: string
          nullable: true
          example: test_user
        action:
          type: string
          example: opened
        created_at:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
        text:
          type: string
          nullable: true
          example: This is a test note
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 250
        total_pages:
          type: integer
          example: 13
    CursorPagination:
      type: object
      description: Cursor-based (keyset) pagination; used when request includes after parameter
      properties:
        limit:
          type: integer
          example: 20
        next_cursor:
          type: string
          description: Opaque cursor for the next page; omitted when there are no more results
...
paths:
...
  /notes-api/v1/notes:
    get:
      summary: Search notes with filters
      tags: [Notes]
      description: |
        Basic search supports standard filters. Advanced search is enabled when
        text or operator parameters are provided (text search in comments, AND/OR).
      security:
        - UserAgent: []
      parameters:
        - in: query
          name: country
          schema:
            type: integer
          description: Filter by country ID
        - in: query
          name: status
          schema:
            type: string
            enum: [open, closed, reopened]
          description: Filter by status
        - in: query
          name: user_id
          schema:
            type: integer
          description: Filter by user ID
        - in: query
          name: date_from
          schema:
            type: string
            format: date
          description: Filter notes created from this date (YYYY-MM-DD)
        - in: query
          name: date_to
          schema:
            type: string
            format: date
          description: Filter notes created until this date (YYYY-MM-DD)
        - in: query
          name: bbox
          schema:
            type: string
          description: Bounding box (min_lon,min_lat,max_lon,max_lat)
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Results per page
        - in: query
          name: after
          schema:
            type: string
          description:
            Opaque cursor for keyset pagination (from previous response next_cursor). When set,
            cursor-based pagination is used and page is ignored.
        - in: query
          name: text
          schema:
            type: string
            minLength: 1
            maxLength: 500
          description: Search text in note comments (enables advanced search)
        - in: query
          name: operator
          schema:
            type: string
            enum: [AND, OR]
            default: AND
          description: Logical operator to combine filters
      responses:
        '200':
          description: |
            Search results with pagination. Without `after`, pagination uses page/offset (page, limit, total, total_pages).
            With `after`, cursor-based pagination is used (limit, next_cursor; next_cursor omitted when no more results).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Note'
                  pagination:
                    oneOf:
                      - $ref: '#/components/schemas/Pagination'
                      - $ref: '#/components/schemas/CursorPagination'
                  filters:
                    type: object
        '400':
          description: Invalid parameters (e.g. invalid after cursor)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes-api/v1/notes/{note_id}:
    get:
      summary: Get a note by ID
      tags: [Notes]
      security:
        - UserAgent: []
      parameters:
        - in: path
          name: note_id
          required: true
          schema:
            type: integer
          description: OSM note ID
      responses:
        '200':
          description: Note details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Note'
        '400':
          description: Invalid note ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes-api/v1/notes/{note_id}/comments:
    get:
      summary: Get comments for a note
      tags: [Notes]
      security:
        - UserAgent: []
      parameters:
        - in: path
          name: note_id
          required: true
          schema:
            type: integer
          description: OSM note ID
      responses:
        '200':
          description: List of note comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NoteComment'
                  count:
                    type: integer
                    example: 2
        '400':
          description: Invalid note ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...